Added feature add-reviews fix3+

+++ src/main/java/ru/yandex/practicum/filmorate/controller/ReviewController.java
@@ -33,7 +33,9 @@ public class ReviewController {
 
     @DeleteMapping("/{id}")
     public void deleteReview(@PathVariable Long id) {
-        reviewService.deleteReview(id);
+        if (id != null) {
+            reviewService.deleteReview(id);
+        }
     }
 
     @GetMapping("/{id}")
